# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  files.ErrorResponse:
    properties:
      code:
        type: string
      details:
        type: string
      error:
        type: string
      success:
        type: boolean
    type: object
  files.GenerateDownloadURLRequest:
    properties:
      file_key:
        type: string
    required:
    - file_key
    type: object
  files.GenerateDownloadURLResponse:
    properties:
      download_url:
        type: string
      expires_at:
        description: Unix timestamp
        type: integer
    type: object
  files.GenerateUploadURLRequest:
    properties:
      content_type:
        type: string
      filename:
        type: string
      max_size:
        description: 'Optional: max file size in bytes'
        type: integer
    required:
    - content_type
    - filename
    type: object
  files.GenerateUploadURLResponse:
    properties:
      expires_at:
        description: Unix timestamp
        type: integer
      file_key:
        type: string
      upload_url:
        type: string
    type: object
  follow.ErrorResponse:
    properties:
      error:
        type: string
      success:
        type: boolean
    type: object
  follow.FollowRequest:
    properties:
      target_user_id:
        type: string
    required:
    - target_user_id
    type: object
  follow.FollowResponse:
    properties:
      message:
        type: string
      success:
        type: boolean
    type: object
  likes.CountResponse:
    properties:
      count:
        type: integer
      post_id:
        type: string
    type: object
  likes.Like:
    properties:
      created_at:
        type: string
      id:
        type: string
      post_id:
        type: string
      user_id:
        type: string
    type: object
  likes.LikeRequest:
    properties:
      post_id:
        type: string
    required:
    - post_id
    type: object
  likes.LikedResponse:
    properties:
      liked:
        type: boolean
      post_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@instant.com
    name: API Support
  description: |-
    Microservices API Gateway for Instant social media platform
    This API provides access to authentication, posts, files, likes, comments, and follow services.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Instant API Gateway
  version: "1.0"
paths:
...
  /api/files/{key}:
    delete:
      description: Delete a file from S3/MinIO storage (requires authentication)
      parameters:
      - description: File key
        in: path
        name: key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/files.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/files.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Delete a file
      tags:
      - files
  /api/files/download-url:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for downloading a file from S3/MinIO
      parameters:
      - description: File download request
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/files.GenerateDownloadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/files.GenerateDownloadURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/files.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/files.ErrorResponse'
      summary: Generate presigned download URL
      tags:
      - files
  /api/files/upload-url:
    post:
      consumes:
      - application/json
      description: Generate a presigned URL for uploading a file to S3/MinIO (requires
        authentication)
      parameters:
      - description: File upload request
        in: body
        name: file
        required: true
        schema:
          $ref: '#/definitions/files.GenerateUploadURLRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/files.GenerateUploadURLResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/files.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Generate presigned upload URL
      tags:
      - files
  /api/follow:
    post:
      consumes:
      - application/json
      description: Follow another user by providing their ID (requires authentication)
      parameters:
      - description: Follow request data
        in: body
        name: follow
        required: true
        schema:
          $ref: '#/definitions/follow.FollowRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/follow.FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Follow a user
      tags:
      - follow
  /api/follow/{user_id}:
    delete:
      description: Unfollow a user by their ID (requires authentication)
      parameters:
      - description: User ID (UUID) to unfollow
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/follow.FollowResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
      security:
      - SessionAuth: []
      summary: Unfollow a user
      tags:
      - follow
  /api/follow/{user_id}/followers:
    get:
      description: Retrieve the list of users who follow the specified user
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
      summary: Get user's followers
      tags:
      - follow
  /api/follow/{user_id}/following:
    get:
      description: Retrieve the list of users that the specified user is following
      parameters:
      - description: User ID (UUID)
        in: path
        name: user_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/follow.ErrorResponse'
      summary: Get users being followed
      tags:
      - follow
  /api/likes:
    post:
      consumes:
      - application/json
      description: Like a post by providing post ID (requires authentication)
      parameters:
      - description: Like request data
        in: body
        name: like
        required: true
        schema:
          $ref: '#/definitions/likes.LikeRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/likes.Like'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Like a post
      tags:
      - likes
  /api/likes/{post_id}:
    delete:
      description: Remove like from a post (requires authentication)
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Unlike a post
      tags:
      - likes
...
  /api/posts/{post_id}/likes/count:
    get:
      description: Get the total number of likes for a specific post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/likes.CountResponse'
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get like count for a post
      tags:
      - likes
  /api/posts/{post_id}/likes/me:
    get:
      description: Check if the authenticated user has liked a specific post
      parameters:
      - description: Post ID
        in: path
        name: post_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/likes.LikedResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - SessionAuth: []
      summary: Check if current user liked a post
      tags:
      - likes
...
securityDefinitions:
  SessionAuth:
    description: Session-based authentication via cookie
    in: cookie
    name: session_id
    type: apiKey
swagger: "2.0"